1) Put code from specific/android part to shared folder 2) Create TODOs 3) Refactoring
