+++ androidApp/src/main/java/com/serj/recommend/android/ui/screens/main/search/SearchScreen.kt
@@ -22,7 +22,8 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.serj.recommend.android.R
-import com.serj.recommend.android.model.RecommendationType
+import com.serj.recommend.android.model.getColor
+import com.serj.recommend.datalayer.RecommendationType
 
 
 @Composable
