+++ gradle/libs.versions.toml
@@ -6,6 +6,7 @@ compose-compiler = "1.5.6"
 compose-material3 = "1.1.2"
 androidx-activityCompose = "1.8.2"
 appcompat = "1.6.1"
+kotlinxDatetime = "0.5.0"
 material = "1.11.0"
 
 [libraries]
@@ -18,6 +19,7 @@ compose-foundation = { module = "androidx.compose.foundation:foundation", versio
 compose-material3 = { module = "androidx.compose.material3:material3", version.ref = "compose-material3" }
 androidx-navigation-runtime-ktx = { group = "androidx.navigation", name = "navigation-runtime-ktx", version = "2.7.6" }
 appcompat = { group = "androidx.appcompat", name = "appcompat", version.ref = "appcompat" }
+kotlinx-datetime = { module = "org.jetbrains.kotlinx:kotlinx-datetime", version.ref = "kotlinxDatetime" }
 material = { group = "com.google.android.material", name = "material", version.ref = "material" }
 androidx-media3-common = { group = "androidx.media3", name = "media3-common", version = "1.2.0" }
 firebase-auth-ktx = { group = "com.google.firebase", name = "firebase-auth-ktx", version = "22.3.0" }
