+++ shared/build.gradle.kts
@@ -29,6 +29,7 @@ kotlin {
         val commonMain by getting {
             dependencies {
                 //put your multiplatform dependencies here
+                implementation(libs.kotlinx.datetime)
             }
         }
         val commonTest by getting {
