+++ shared/src/commonMain/kotlin/com/serj/recommend/datalayer/RecommendationType.kt
@@ -1,0 +1,6 @@
+package com.serj.recommend.datalayer
+
+enum class RecommendationType {
+    Film, Music, Book,
+    Series, Perfume, Place
+}
